No changes, just the script classes being organized in their own source and header files
